requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
click==7.1.2
Django==3.0.7
djangorestframework==3.11.0
Flask==1.1.2
idna==2.10
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
pyee==7.0.2
pytz==2020.1
slackclient==2.7.2
slackeventsapi==2.2.0
sqlparse==0.3.1
typing-extensions==3.7.4.2
Werkzeug==1.0.1
yarl==1.4.2
